Adding fetch-size parameter for the common DB driver definition
Changing execute.clj to set the fetch size in the PreparedStatement it creates, if present in the driver

Adding test for all drivers for which we're adding the parameter that assert that the ResultSet is produced with the given fetchSize

Moving test query (i.e. "SELECT 1") to new multimethod, and overriding for Oracle to be "SELECT 1 FROM DUAL"

Replacing some places where test query was hard coded with new multimethod invocation

Adding fetch-size connection property to the following drivers: Postgres, Redshift, MySQL, Oracle

Adding driver as param for can-connect-with-spec? so that the test query can be looked up from the multimethod